backend2/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.11.7

# CORS middleware
python-multipart==0.0.6

# AI Provider Dependencies (uncomment as needed)
# OpenAI
openai==1.3.0

# Anthropic Claude
# anthropic==0.7.0

# Google Gemini
# google-generativeai==0.3.0

# Local LLM via Ollama
# aiohttp==3.9.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0

# Environment management
python-dotenv==1.0.0

# Logging
structlog==23.2.0
